remove /etc/xCATMN if exists
git-svn-id: https://svn.code.sf.net/p/xcat/code/xcat-core/trunk@2521 8638fb3e-16cb-4fca-ae20-7b5d299a9bcd

+++ xCATsn/xCATsn.spec
@@ -52,8 +52,9 @@ if [ "$1" = "1" ]; then #Only if installing for the first time..
 # makes it a service node
   touch /etc/xCATSN
 # remove any management node file
+if [ -f /etc/xCATMN ]; then
   rm  /etc/xCATMN
-
+fi
 
 
 ###  Start the xcatd daemon
